Multiplexing seems to be in working order now. Some tests are still needed to be carried out. The code still needs to be cleaned a bit.

+++ drm4g/communicators/openssh.py
@@ -68,7 +68,7 @@ class Communicator(drm4g.communicators.Communicator):
 
 
     def createConfFiles(self):
-        logger.debug("\nRunning createConfFiles function\n")
+        logger.debug("\nRunning createConfFiles function from "+self.parent_module+"\n")
         #the maximum length of the path of a unix domain socket is 108 on Linux, 104 on Mac OS X
         conf_text = ("Host *\n"
             "    ControlMaster auto\n"
@@ -81,6 +81,7 @@ class Communicator(drm4g.communicators.Communicator):
         if not exists(Communicator.socket_dir):
             logger.debug("\nCreating socket directory in "+Communicator.socket_dir+".\n")
             subprocess.call('mkdir -p %s' % (Communicator.socket_dir), shell=True)
+        logger.debug("\nEnding createConfFiles function from "+self.parent_module+"\n")
 
     def connect(self):
         """
@@ -90,45 +91,60 @@ class Communicator(drm4g.communicators.Communicator):
             p_module=sys._getframe().f_back.f_code.co_filename
             p_function=sys._getframe().f_back.f_code.co_name
             logger.debug("\nRunning connect function\n - called from "+p_module+"\n - by function "+p_function+"\n")
+            
+            if not self.configfile:
+                logger.debug("\nself.configfile is not defined\n")
+                raise Exception("\n\nself.configfile is not defined\n\n")
+
             if not exists(self.configfile):
                 self.createConfFiles()
             logger.debug("\nThe socket "+join(Communicator.socket_dir, '%s-%s@%s:%s' % (self.parent_module ,self.username, self.frontend, self.port))+" existance is "+str(exists(join(Communicator.socket_dir, '%s-%s@%s:%s' % (self.parent_module ,self.username, self.frontend, self.port))))+".\n")
 
             if not exists(join(Communicator.socket_dir, '%s-%s@%s:%s' % (self.parent_module ,self.username, self.frontend, self.port))):
-                logger.debug("\nNo socket exists for "+self.parent_module+" so a new one will be created.\n")
+                logger.debug("\nNo master conncection exists for "+self.parent_module+" so a new one will be created.\n")
                 def first_ssh():
-                    logger.debug("\nCreating first connection for "+self.parent_module+"\n")
-                    command = 'ssh -F %s -i %s -p %s %s@%s' % (self.configfile, self.private_key, str(self.port), self.username, self.frontend)
+                    logger.debug("\nRunning first_ssh function\n - Creating first connection for "+self.parent_module+"\n")
+                    #this is here because the threads are created at the same time, so the moent one creates the conection, the rest are going to cause an exception (which i'm not really sure why since ContloMaster is set to auto)
+                    if not exists(join(Communicator.socket_dir, '%s-%s@%s:%s' % (self.parent_module ,self.username, self.frontend, self.port))):
+                        command = 'ssh -F %s -i %s -p %s -T %s@%s' % (self.configfile, self.private_key, str(self.port), self.username, self.frontend)
                     pipe = subprocess.Popen(command.split(), stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-                    #si quiero usar este msj de logger, hacer que el comando se ejecute en background
-                    #logger.debug("\nDone creating first connection for "+self.parent_module+"\n")
                     out,err = pipe.communicate()
 
-                    if "too long for Unix domain socket" in str(err):
-                        logger.debug("\nSocket path was too long for Unix domain socket.\nCreating sockets in ~/.ssh/dmr4g.\nException captured in first_ssh.\n")
-                        #subprocess.call('rmdir %s' % (self.socket_dir), shell=True)
-                        Communicator.socket_dir = join(expanduser('~'), '.ssh/drm4g')
-                        self.createConfFiles()
-                        logger.debug("\nCalling first_ssh once again, but with a new socket_dir\n")
-                        first_ssh()
-                    elif "disabling multiplexing" in str(err):
-                        logger.debug("\nMux isn't working from the connect function. Eliminating "+self.parent_module+"'s socket file.\n This shouldn't be appearing since it's supposedly the first time that it's being created.\n")
-                        subprocess.call("rm -r %s/%s-%s@%s:%s" % (Communicator.socket_dir,self.parent_module,self.username,self.frontend,str(self.port)),shell=True)
-                        raise Exception(str(err))
-                    else:
-                        raise Exception(str(err))
+                    if err:
+                        if "too long for Unix domain socket" in str(err):
+                            logger.debug("\nSocket path was too long for Unix domain socket.\nCreating sockets in ~/.ssh/dmr4g.\nException captured in first_ssh.\n")
+                            Communicator.socket_dir = join(expanduser('~'), '.ssh/drm4g')
+                            self.createConfFiles()
+                            logger.debug("\nCalling first_ssh once again, but with a new socket_dir\n")
+                            first_ssh()
+                        elif "disabling multiplexing" in str(err):
+                            logger.debug("\nMux isn't working from the connect function. Eliminating "+self.parent_module+"'s socket file.\n This shouldn't be appearing since it's supposedly the first time that it's being created.\n")
+                            subprocess.call("rm -r %s/%s-%s@%s:%s" % (Communicator.socket_dir,self.parent_module,self.username,self.frontend,str(self.port)),shell=True)
+                            first_ssh()
+                            #raise Exception(str(err))
+                        else:
+                            logger.debug("\nUnexpected error occured while running first_ssh:\n"+str(err))
+                            raise Exception(str(err))
 
                 t = threading.Thread(target=first_ssh, args = ())
                 t.daemon = True
                 logger.debug("\nStarting thread with first_ssh\n")
                 t.start()
-                time.sleep(5)
-            
+                time.sleep(5) #so that there's time to make the first connection in case there was an error
+                #cont=0
+                # while not exists(join(Communicator.socket_dir, '%s-%s@%s:%s' % (self.parent_module ,self.username, self.frontend, self.port))) and cont < 130:
+                #     cont+=1
+                #     time.sleep(1)
+                # if not cont < 130:
+                #     logger.debug("\n**Tried to wait for first connection to be established, but after 2 minutes still nothing\n")
+
             if self.conn==None:
                 logger.debug("\nNo conn exists (conn == "+str(self.conn)+") for "+self.parent_module+" so a new one will be created.\n")
                 self.conn = SSHConnection(self.frontend, login=self.username, port=str(self.port),
                     configfile=self.configfile, identity_file=self.private_key,
                     ssh_agent_socket=self.agent_socket, timeout=SSH_CONNECT_TIMEOUT)
+
+            logger.debug("\nEnding connect function from "+self.parent_module+"\n")
                 
         except Exception as error :
             if "too long for Unix domain socket" in str(error):
@@ -140,40 +156,38 @@ class Communicator(drm4g.communicators.Communicator):
                 raise
 
     def execCommand(self , command , input = None ):
+        """
+        parent_module: /home/antonio/virtualenv_drm4g/3_drm4g_mux/local/lib/python2.7/site-packages/drm4g/managers/__init__.py
+        parent_function: system_information
+        dirname parent_module: /home/antonio/virtualenv_drm4g/3_drm4g_mux/local/lib/python2.7/site-packages/drm4g/managers
+        basename parent_module: __init__.py
+        """
         try:
-            logger.debug("\nRunning execCommand function\n")
-            '''parent_module=sys._getframe().f_back.f_code.co_filename
-            parent_function=sys._getframe().f_back.f_code.co_name'''
-            logger.debug("\nTrying to execute command "+str(command)+" in function execCommand.\n")
-            logger.debug("\nGoing to run connect function.\nThat should already have been done, so it shouldn't do anything.\n")
-            self.connect()
-            logger.debug("\nFinnished running connect function.\n")
-            #if command == 'uname -n -r -m -o':
-            #    raise Exception("\n\nMy configfile is: "+self.configfile+"\nMy socket_dir is: "+self.socket_dir+"\n\n")
+            logger.debug("\nRunning execCommand function from "+self.parent_module+"\n    - Trying to execute command "+str(command)+"\n")
+            
+            if not self.conn:
+                logger.debug("\nGoing to run connect function.\n - That should already have been done, so it shouldn't do anything.\n")
+                self.connect()
+            
             ret = self.conn.run(command)
+            logger.debug("\nEnding execCommand function.\n")
             return ret.stdout , ret.stderr
         except Exception as excep:
             if "disabling multiplexing" in str(excep):
                 logger.debug("\nMux isn't working from the execCommand function. Eliminating "+self.parent_module+"'s socket file.\n")
                 subprocess.call("rm -r %s/%s-%s@%s:%s" % (Communicator.socket_dir,self.parent_module,self.username,self.frontend,str(self.port)),shell=True)
                 self.execCommand(command, input)
-            #elif "too long for Unix domain socket" in str(excep):
-            #    self.socket_dir = join(expanduser('~'), '.ssh/drm4g')
-            #    self.createConfFiles('im')
-            #    self.createConfFiles('tm')
-            #    self.createConfFiles('em')
-            #    self.connect()
             else:
                 raise
 
     def mkDirectory(self, url):
         try:
-            logger.debug("\nRunning mkDirectory function\n")
-            self.connect()
+            logger.debug("\nRunning mkDirectory function from "+self.parent_module+"\n")
             to_dir         = self._set_dir(urlparse(url).path)
             stdout, stderr = self.execCommand( "mkdir -p %s" % to_dir )
             if stderr :
                 raise ComException( "Could not create %s directory: %s" % ( to_dir , stderr ) )
+            logger.debug("\nEnding mkDirectory function from "+self.parent_module+"\n")
         except Exception as excep:
             if "disabling multiplexing" in str(excep):
                 logger.debug("\nMux isn't working from the mkDirectory function. Eliminating "+self.parent_module+"'s socket file.\n")
@@ -185,12 +199,12 @@ class Communicator(drm4g.communicators.Communicator):
 
     def rmDirectory(self, url):
         try:
-            logger.debug("\nRunning rmDirectory function\n")
-            self.connect()
+            logger.debug("\nRunning rmDirectory function from "+self.parent_module+"\n")
             to_dir         = self._set_dir(urlparse(url).path)
             stdout, stderr = self.execCommand( "rm -rf %s" % to_dir )
             if stderr:
                 raise ComException( "Could not remove %s directory: %s" % ( to_dir , stderr ) )
+            logger.debug("\nEnding rmDirectory function from "+self.parent_module+"\n")
         except Exception as excep:
             if "disabling multiplexing" in str(excep):
                 logger.debug("\nMux isn't working from the rmDirectory function. Eliminating "+self.parent_module+"'s socket file.\n")
@@ -203,8 +217,9 @@ class Communicator(drm4g.communicators.Communicator):
 
     def copy( self , source_url , destination_url , execution_mode = '' ) :
         try:
-            logger.debug("\nRunning copy function\n")
-            self.connect()
+            logger.debug("\nRunning copy function from "+self.parent_module+"\n")
+            if not self.conn: 
+                self.connect()
             with self._sem :
                 if 'file://' in source_url :
                     from_dir = urlparse( source_url ).path
@@ -218,6 +233,7 @@ class Communicator(drm4g.communicators.Communicator):
                     from_dir = self._set_dir( urlparse( source_url ).path )
                     to_dir   = urlparse(destination_url).path
                     self.remote_scp( [from_dir] , target=to_dir )
+            logger.debug("\nEnding copy function from "+self.parent_module+"\n")
         except Exception as excep:
             if "disabling multiplexing" in str(excep):
                 logger.debug("\nMux isn't working from the copy function. Eliminating "+self.parent_module+"'s socket file.\n")
@@ -229,13 +245,14 @@ class Communicator(drm4g.communicators.Communicator):
 
     #internal
     def _set_dir(self, path):
-        logger.debug("\nRunning _set_dir function\n")
+        logger.debug("\nRunning _set_dir function from "+self.parent_module+"\n")
         work_directory =  re.compile( r'^~' ).sub( self.work_directory , path )
+        logger.debug("\nEnding _set_dir function from "+self.parent_module+"\n")
         return  work_directory
 
 
     def remote_scp(self, files, target):
-        logger.debug("\nRunning remote_scp function\n")
+        logger.debug("\nRunning remote_scp function from "+self.parent_module+"\n")
         scp_command = self.scp_command(files, target)
         pipe = subprocess.Popen(scp_command,
                 stdin=subprocess.PIPE, stdout=subprocess.PIPE,
@@ -254,13 +271,14 @@ class Communicator(drm4g.communicators.Communicator):
         returncode = pipe.returncode
         if returncode != 0:  # ssh client error
             raise ComException("%s (under %s): %s" % (' '.join(scp_command), self.username, err.strip()))
+        logger.debug("\nEnding remote_scp function from "+self.parent_module+"\n")
 
     def scp_command(self, files, target, debug=False):
         """
         Build the command string to transfer the files identified by filenames.
         Include target(s) if specified. Internal function
         """
-        logger.debug("\nRunning scp_command function\n")
+        logger.debug("\nRunning scp_command function from "+self.parent_module+"\n")
         cmd = ['scp', debug and '-vvvv' or '-q', '-r']
 
         if self.username:
@@ -283,6 +301,7 @@ class Communicator(drm4g.communicators.Communicator):
             cmd.append('%s:%s' % (remotename, f))
         cmd.append(target)
         logger.debug("\nThe command is "+str(cmd)+"\n")
+        logger.debug("\nEnding scp_command function from "+self.parent_module+"\n")
         return cmd
 
     def get_env(self):
@@ -290,8 +309,9 @@ class Communicator(drm4g.communicators.Communicator):
         Retrieve environment variables and replace SSH_AUTH_SOCK
         if ssh_agent_socket was specified on object creation.
         """
-        logger.debug("\nRunning get_env function\n")
+        logger.debug("\nRunning get_env function from "+self.parent_module+"\n")
         env = os.environ.copy()
         if self.agent_socket: #should i check that it's empty? "if not env['SSH_AUTH_SOCK'] and self.agent_socket:"
             env['SSH_AUTH_SOCK'] = self.agent_socket
+        logger.debug("\nEnding get_env function from "+self.parent_module+"\n")
         return env
